pyproject.toml
[project]
name = "rejseplanen-cli"
version = "0.1.0"
description = "Modern CLI tool for Danish public transport using Rejseplanen API 2.0"
authors = [
    {name = "Andrea Di Luca", email = "andreadiluca10@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "click>=8.1.0",
    "requests>=2.31.0",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "keyring>=24.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "types-requests>=2.31.0",
]

[project.scripts]
rejs = "rejseplanen.cli:cli"

[project.urls]
Homepage = "https://github.com/andreadiluca/rejseplanen-cli"
Repository = "https://github.com/andreadiluca/rejseplanen-cli"
Issues = "https://github.com/andreadiluca/rejseplanen-cli/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rejseplanen"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise from err (not critical for this app)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]